second commit

+++ routes/events.js
@@ -4,57 +4,83 @@ const router = Router({ mergeParams: true });
 
 const EventsDAO = require('../daos/events');
 
+const CalendarDAO = require('../daos/calendars');
+
 //const router = Router();
 
 router.get('/', async (req, res, next) => {
-  try {
-    const events = await EventsDAO.getAll();
-    if (events) {
-      res.json(events);
-    } else {
-      res.sendStatus(404);
-    }
-  } catch (e) {
+  const { calendarId } = req.params;
+  const id = req.params.id;
+  console.log(calendarId);
+  const selectedCalendar = await CalendarDAO.getById(calendarId);
+  console.log(selectedCalendar);
+  if (!selectedCalendar) {
+    console.log('Sending 404 ');
     res.sendStatus(404);
-    next(e);
+  } else {
+    try {
+      const events = await EventsDAO.getAll();
+      // if (events) {
+      res.json(events);
+      // } else {
+      //   res.sendStatus(404);c
+      // }
+    } catch (e) {
+      // res.sendStatus(404);
+      next(e);
+    }
   }
 });
 
 router.get('/:id', async (req, res, next) => {
   const { calendarId } = req.params;
   const id = req.params.id;
-
-  try {
-    const selectedEvent = await EventsDAO.getById(id);
-    if (selectedEvent) {
-      res.json(selectedEvent);
-    } else {
-      res.sendStatus(404);
-    }
-  } catch (e) {
+  console.log(calendarId);
+  const selectedCalendar = await CalendarDAO.getById(calendarId);
+  console.log(selectedCalendar);
+  if (!selectedCalendar) {
+    console.log('Sending 404 ');
     res.sendStatus(404);
-    next(e);
+  } else {
+    try {
+      const selectedEvent = await EventsDAO.getById(id, calendarId);
+      if (selectedEvent) {
+        res.json(selectedEvent);
+      } else {
+        res.sendStatus(404);
+      }
+    } catch (e) {
+      res.sendStatus(404);
+      next(e);
+    }
   }
 });
 
 router.post('/', async (req, res, next) => {
   const { calendarId } = req.params;
   console.log(calendarId);
-
-  const { name } = req.body;
-  if (!name || JSON.stringify(name) === '{}') {
-    res.status(400).send('Name is required and can not be Blank');
+  const selectedCalendar = await CalendarDAO.getById(calendarId);
+  console.log(selectedCalendar);
+  if (!selectedCalendar) {
+    console.log('Sending 404 ');
+    res.sendStatus(404);
   } else {
-    try {
-      const SelectedEvent = await EventsDAO.create(req.body, calendarId);
-      if (SelectedEvent) {
-        res.json(SelectedEvent);
-      } else {
+    const { name } = req.body;
+    if (!name || JSON.stringify(name) === '{}') {
+      res.status(400).send('Name is required and can not be Blank');
+    } else {
+      try {
+        console.log('On Post req.  inside of eventSelect ');
+        const SelectedEvent = await EventsDAO.create(req.body, calendarId);
+        if (SelectedEvent) {
+          res.json(SelectedEvent);
+        } else {
+          res.sendStatus(400);
+        }
+      } catch (e) {
         res.sendStatus(400);
+        next(e);
       }
-    } catch (e) {
-      res.sendStatus(400);
-      next(e);
     }
   }
 });
@@ -69,7 +95,11 @@ router.put('/:id', async (req, res, next) => {
     res.status(400).send('Event name and Date  are required');
   } else {
     try {
-      const updatedEvent = await EventsDAO.updateById(req.params.id, req.body);
+      const updatedEvent = await EventsDAO.updateById(
+        req.params.id,
+        calendarId,
+        req.body
+      );
       if (updatedEvent) {
         res.json(updatedEvent);
       } else {
@@ -84,21 +114,28 @@ router.put('/:id', async (req, res, next) => {
 
 router.delete('/:id', async (req, res, next) => {
   const { calendarId } = req.params;
-  console.log(calendarId);
-  const id = req.params.id;
-  console.log(id);
-  try {
-    const event = await EventsDAO.removeById(id);
-
-    if (event) {
-      res.json(event);
-      // res.sendStatus(200);
-    } else {
-      res.sendStatus(404);
-    }
-  } catch (e) {
+  //console.log(calendarId);
+  const selectedCalendar = await CalendarDAO.getById(calendarId);
+  //console.log(selectedCalendar);
+  if (!selectedCalendar) {
+    console.log('Sending 404 ');
     res.sendStatus(404);
-    next(e);
+  } else {
+    const id = req.params.id;
+
+    try {
+      const event = await EventsDAO.removeById(id);
+
+      if (event) {
+        res.json(event);
+        // res.sendStatus(200);
+      } else {
+        res.sendStatus(404);
+      }
+    } catch (e) {
+      res.sendStatus(404);
+      next(e);
+    }
   }
 });
